+++ src/main/java/prefuse/util/force/EulerIntegrator.java
@@ -16,18 +16,18 @@ public class EulerIntegrator implements Integrator {
      * @see prefuse.util.force.Integrator#integrate(prefuse.util.force.ForceSimulator, long)
      */
     public void integrate(ForceSimulator sim, long timestep) {
-        float speedLimit = sim.getSpeedLimit();
+        double speedLimit = sim.getSpeedLimit();
         Iterator iter = sim.getItems();
         while ( iter.hasNext() ) {
             ForceItem item = (ForceItem)iter.next();
             item.location[0] += timestep * item.velocity[0];
             item.location[1] += timestep * item.velocity[1];
-            float coeff = timestep / item.mass;
+            double coeff = timestep / item.mass;
             item.velocity[0] += coeff * item.force[0];
             item.velocity[1] += coeff * item.force[1];
-            float vx = item.velocity[0];
-            float vy = item.velocity[1];
-            float v = (float)Math.sqrt(vx*vx+vy*vy);
+            double vx = item.velocity[0];
+            double vy = item.velocity[1];
+            double v = (double)Math.sqrt(vx*vx+vy*vy);
             if ( v > speedLimit ) {
                 item.velocity[0] = speedLimit * vx / v;
                 item.velocity[1] = speedLimit * vy / v;
